Keep chat identity imports narrow
